* MDF [parquet] change key from uint32 to uint64(header).

+++ include/nng/supplemental/nanolib/parquet.h
@@ -9,7 +9,7 @@ extern "C" {
 #endif
 
 typedef struct {
-	uint32_t *keys;
+	uint64_t *keys;
 	uint8_t **darray;
 	uint32_t *dsize;
 	uint32_t  size;
@@ -17,13 +17,13 @@ typedef struct {
 	void     *arg;
 } parquet_object;
 
-parquet_object *parquet_object_alloc(uint32_t *keys, uint8_t **darray,
+parquet_object *parquet_object_alloc(uint64_t *keys, uint8_t **darray,
     uint32_t *dsize, uint32_t size, nng_aio *aio, void *arg);
 int             parquet_write_batch_async(parquet_object *elem);
 int             parquet_write_launcher(conf_parquet *conf);
 
-const char  *parquet_find(uint32_t key);
-const char **parquet_find_span(uint32_t key, uint32_t offset, uint32_t *size);
+const char  *parquet_find(uint64_t key);
+const char **parquet_find_span(uint64_t key, uint32_t offset, uint32_t *size);
 
 #ifdef __cplusplus
 }
